Must provide relative path to detected dir and its creation
Else the PolyDiskTree accesses the newly created Directory which does
not have a rP to sort it in with.

We provide parent_id and relative_path on creation of the new child.
Until now, this would mean the relative_path to be from the parent
directory, but this would complicate code extremely, so we just accept
both together.